Refactored ready for testing
